Alter the repository structural layout for the BSR
Our next big modification to the help system is to flip the layout of
everything so that instead of being a library that packs in a package,
instead we will have a package that packs in the library instead.

User code should still be able to access the API via commands; we will
attempt to not publicly expose it directly to protect ourselves from update
issues.