Fix a silly keybinding syntax issue
There was an issue with the syntax definition on key bindings in which
we thought it was not correctly scoping the < character. It turns out
that it was actually not correctly scopingthe > character instead.

This fixes that, although it now makes both < and > highlight differently
than the key binding itself.

It is, at least, consistent now.